# backend/documentation/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a PIXAGRAM-API documentation server.",
    "version": "1.0.0",
    "title": "PIXAGRAM-API Documentation",
    "termsOfService": "",
    "contact": {
      "email": "kanika.nadkarni9@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/post/{id}": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Get post by id",
        "description": "",
        "operationId": "getPost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The post will be fetched by id.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "404": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "post"
        ],
        "summary": "edit a post",
        "description": "",
        "operationId": "editPost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The post will be edited by id.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "caption",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "caption": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "delete a post",
        "description": "",
        "operationId": "deletePost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The post will be edited by id.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/post/user/{email}": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Find posts realted to user by user login-email",
        "description": "Returns all posts related to user",
        "operationId": "getPostOfUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "email of user to return all posts",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "posts not found"
          }
        }
      },
      "post": {
        "tags": [
          "post"
        ],
        "summary": "Insert new post with image for a logged-in user with parameter as his login-email",
        "description": "",
        "operationId": "createPostofUser",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "logged-in email of an user",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/NewPost"
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
    "/api/post/like/{id}": {
      "post": {
        "tags": [
          "post"
        ],
        "summary": "like a post",
        "description": "",
        "operationId": "likePost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The post will be fetched by id.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "email in body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/post/unlike/{id}": {
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "unlike a post",
        "description": "",
        "operationId": "unlikePost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id as parameter.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "email in body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "postunlike": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/post/comment/{id}": {
      "post": {
        "tags": [
          "post"
        ],
        "summary": "add a comment",
        "description": "",
        "operationId": "addComment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The post will be edited by id.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "object in body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "c_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "post"
        ],
        "summary": "edit a comment",
        "description": "",
        "operationId": "editComment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id for a post.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "caption",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "delete a comment",
        "description": "",
        "operationId": "deleteComment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "comment id.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/post/comments/{id}": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Getting all comments for a post",
        "description": "",
        "operationId": "getComments",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post-id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/Comments"
            }
          },
          "404": {
            "description": "Warning",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user e-mail",
        "description": "",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The user will be fetched by user e-mail.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Post": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "Post": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "postTimestamp": {
              "type": "string",
              "format": "date-time"
            },
            "imagePath": {
              "type": "string"
            },
            "caption": {
              "type": "string"
            },
            "userEmail": {
              "type": "string"
            }
          }
        }
      }
    },
    "Posts": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "postTimestamp": {
                "type": "string",
                "format": "date-time"
              },
              "imagePath": {
                "type": "string"
              },
              "caption": {
                "type": "string"
              },
              "userEmail": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "NewPost": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "imagePath": {
                "type": "string"
              },
              "caption": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "Category"
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "pswd": {
          "type": "string"
        },
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "profileimagePath": {
          "type": "string"
        },
        "coverimagePath": {
          "type": "string"
        },
        "visibility": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
    "Comments": {
      "type": "object",
      "properties": {
        "fname": {
          "type": "string"
        },
        "lname": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "commentsTimeStamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}