# backend/routes/post.js
const express = require("express");
const router = express.Router();
const multer = require("multer");
//const db = require('../models/database');
const db = require('../models/database');

const MIME_TYPE_MAP = {
  "image/png": "png",
  "image/jpeg": "jpg",
  "image/jpg": "jpg"
};

const storage = multer.diskStorage({
  destination: (req, file, cb) => {
    const isValid = MIME_TYPE_MAP[file.mimetype];
    let error = new Error("Invalid mime type");
    if (isValid) {
      error = null;
    }
    cb(error, "backend/images");
  },
  filename: (req, file, cb) => {
    const name = file.originalname
      .toLowerCase()
      .split(" ")
      .join("-");
    const ext = MIME_TYPE_MAP[file.mimetype];
    cb(null, name + "-" + Date.now() + "." + ext);
  }
});


router.get('/:id', (req,res) => getPost(req,res));
router.get('/user/:email', (req, res) => getPostsOfUser(req,res));
router.post('/user/:email', multer({storage: storage}).single('image'), (req, res) => createPostOfUser(req,res));
router.post('/like/:id' , (req,res) => likePost(req,res));
router.delete('/unlike/:id',(req,res) => unlikePost(req,res));
router.put('/:id',(req,res) => editPost(req,res));
router.post('/comment/:id',(req,res) => addComment(req,res));
router.get('/comments/:id',(req,res) => getComments(req,res));
router.put('/comment/:id',(req,res) => editComment(req,res));
router.delete('comment/:id',(req,res) => deleteComment(req,res));
router.delete('/:id',(req,res) => deletePost(req,res));

function getPostsOfUser(req, res) {
  console.log("Get all posts of a user");
  const sql = 'select * from Posts where userEmail = ?';
  const params = [req.params.email];

  //db.all(sql, params, (err, rows) => {
    db.query(sql, params, (err, rows) => {
    if (err) {
      res.status(404).json({ message: err.message });
      return;
    }
    if(rows.length==0){
      return res.status(401).json({
        message: 'User has no posts yet!'
      })
    }
    res.status(200).json({
      message: 'success',
      photos: rows,
    });
  });
}

function createPostOfUser(req, res) {
  console.log("create a new post");
  const url = req.protocol + '://' + req.get("host");
  const newPhoto = {
    imagePath: url + "/images/" + req.file.filename,
    caption: req.body.caption,
    email: req.params.email
  }
  
  console.log(newPhoto);

  const insert = 'INSERT INTO Posts (imagePath, caption, userEmail) VALUES (?,?,?)';
  const values = [newPhoto.imagePath, newPhoto.caption, newPhoto.email];

  db.query(insert, values, function (err, result) {
    if (err) {
      console.log(err);
      res.status(400).json({ message: err.message });
      return;
    }
    res.json({
      message: 'success',
      post: newPhoto,
      id: this.lastID,
    });
  });
}

function getPost(req,res) {
  console.log("getting one post");
  const sql="select * from Posts where id= ?";
  const params=[req.params.id];
  db.query(sql, params, (err, row) => {
    if (err) {
      res.status(404).json({ message: err.message });
      return;
    }
    if(row.length==0){
      return res.status(401).json({
        message: 'No such Post found!'
      })
    }

    res.json({
      message: 'success',
      post: row,
    });
  });
}

function likePost(req,res){
  console.log("like a post");
  const values=[req.body.email,req.params.id];
  const sql="insert into Likes (email, postId) values (?,?)";
  db.query(sql,values,function(err,result){
    if(err){
      console.log(err);
      res.status(400).json({ message: err.message });
      return;
    }
    res.json({
      message:'success',
      postlike:result
    });
  });
}

function unlikePost(req,res){
  console.log("unlike a post");
  const values=[req.body.email,req.params.id];
  const sql="delete from Likes where email=? and postId=?";
  db.query(sql,values,function(err,result){
    if(err){
      console.log(err);
      res.status(400).json({ message: err.message });
      return;
    }
    res.json({
      message:'success',
      postunlike:values
    });
  });
}

function editPost(req,res){
  console.log("edit a post");
  const values=[req.body.caption,req.params.id];
  const sql="update Posts set caption=? where id=?";
  db.query(sql,values,function(err, result){
    if(err){
      console.log(err);
      res.status(400).json({message:err.message});
      return;
    }
    res.json({
      message:'success',
    });
  });
}

function addComment(req,res){
  console.log("adding a new comment");
  const values=[req.params.id,req.body.email,req.body.content];
  const sql='insert into Comments (post_id, email, content) values (?,?,?)';
  db.query(sql,values,function(err,result){
    if(err){
      console.log(err);
      res.status(400).json({message:err.message});
      return;
    }
    // const id1=result.insertId;
    res.json({
      message:'success',
      c_id:result.insertId
    });
  });
}

function getComments(req,res){
  console.log("getting all comments for a post");
  const values=[req.params.id];
  const sql='select Users.fname, Users.lname, Comments.content, Comments.email, Comments.commentsTimestamp from Comments inner join Users on Comments.email=Users.email where post_id=? order by Comments.commentsTimestamp';
  db.query(sql,values,function(err,result){
    if(err){
      console.log(err);
      res.status(400).json({message:err.message});
      return;
    }
    res.json({
      message:'success',
      comments:result
    });
  });
}
// #########
function editComment(req,res){
  console.log("editing a comment");
  const values=[req.body.content,req.params.id]
  const sql="update Comments set content=? where commentId=?";
  db.query(sql,values,(err,rows) =>{
    if(err){
      res.status(400).json({message:err.message});
      return;
    }
    res.status(200).json({
      message:'success'
    });
  });
}

function deleteComment(req,res){
  console.log("deleting a comment");
  const values=[req.params.id];
  const sql='delete from Comments where commentId=?';
  db.query(sql,values,(err,rows) =>{
    if(err){
      res.status(400).json({message:err.message});
      return;
    }
    res.status(200).json({
      message:'success'
    });
  });
}

function deletePost(req,res){
  console.log("delete a post");
  const values=[req.params.id];
  const sql='delete from Posts where id=?';
  db.query(sql,values,(err,rows) =>{
    if(err){
      res.status(400).json({message:err.message});
      return;
    }
    res.status(200).json({
      message:'success'
    });
  });
}



module.exports = router;